+++ internal/state/firebase_states.go
@@ -6,8 +6,10 @@ import (
 	"github.com/goioc/di"
 	"github.com/joomcode/errorx"
 	"github.com/samber/lo"
+	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"gopkg.in/yaml.v3"
 	"reflect"
 	"strconv"
 )
@@ -57,9 +59,20 @@ func (f *FirebaseStates) GetState(userId int64) (*UserState, error) {
 }
 
 func (f *FirebaseStates) SetState(state *UserState) error {
-	_, err := f.fc.Collection(collection).Doc(strconv.FormatInt(state.UserId, 10)).Set(context.Background(), state)
+	if logrus.IsLevelEnabled(logrus.DebugLevel) {
+		stateYaml, err := yaml.Marshal(state)
+		if err != nil {
+			logrus.WithField("userId", state.UserId).Error("failed to serialize user state")
+		} else {
+			logrus.WithField("state", string(stateYaml)).Debug("saving user state")
+		}
+	}
+
+	wr, err := f.fc.Collection(collection).Doc(strconv.FormatInt(state.UserId, 10)).Set(context.Background(), state)
 	if err != nil {
 		return errorx.EnhanceStackTrace(err, "failed to set user state: userId=%v", state.UserId)
+	} else {
+		logrus.WithField("userId", state.UserId).WithField("updateTime", wr.UpdateTime).Debug("user state saved")
 	}
 
 	return nil